requirements.txt
google-genai
opencv-python
numpy
pydobot
python-dotenv
